Fix 30 bugs from full code review + correct 15 documentation inaccuracies
C core (9 fixes):
- NULL-check transformer forward return in all samplers (crash on OOM/GPU failure)
- NULL-check malloc in load_double_block_weights f32 path; free ff_in on failure
- ensure_work_buffers frees partial allocations before returning -1
- Remove static block_idx shadow variable in DEBUG_DOUBLE_BLOCK path
- Fix integer overflow in iris_image_create calloc (cast to size_t)
- Add stderr warning in iris_linear_nobias_bf16 on malloc failure
- NULL-check normed alloc in zi_final_forward
- ftell() result stored as long in iris_img2img_debug_py (%ld format)
- iris_sample.c NULL checks cover CFG, refs, multi-refs, Heun, ancestral paths

Metal/GPU (11 fixes):
- nil check + NSLog on MTLBuffer allocation at 4 OOM-crash sites
- MTLBuffer ARC release in iris_gpu_tensor_alloc_persistent on struct malloc fail
- bufferB pool slot released correctly when bufferC alloc fails in sgemm_impl
- attention_fused_bf16: add head_dim > 128 guard in kernel
- causal_attention_fused / _bf16: add seq > 512 guard in kernel
- Remove __strong from cBuffers calloc array (ARC violation)
- SDPA graph cache: add scale to sdpa_graph_cache_t and include in key comparison
  for both bf16 and f32 caches; prevents silent graph reuse on scale change
- iris_gpu_linear: replace CPU bias loop (which broke batch mode atomicity by
  committing g_tensor_cmd) with existing g_bias_add_pipeline GPU kernel;
  commit now happens only in the unified non-batch path at function end

Python pipeline (6 fixes):
- data_stager._link_or_copy: use os.path.relpath for relative symlinks (absolute
  symlinks become dangling if volume remounts at different path)
- _atomic_symlink in data_stager + cache_manager: uuid-based temp names (prevents
  race between concurrent callers on same directory)
- pipeline_doctor: fix stager remediation commands (--phase → subcommand form)
- orchestrator._recover_prep_window: read heartbeat mtime for started_at so
  hung-prep timer survives orchestrator restart correctly
- data_explorer --fix: shlex.split instead of cmd.split (handles paths with spaces)
- data_explorer _render_status_html: remove dead self-import block

Trainer (5 fixes):
- _internal_prefix moved outside if-block; fixes NameError when siglip cache
  configured without qwen3/vae dirs
- grad_clip_pct captured before _grad_clip_steps reset; heartbeat now correct
- _compute_val_loss: use_null_image=False so val_loss measures adapter conditioning
  rather than forced null-image base reconstruction (TODO: load real siglip feats)
- Checkpoints and JSON sidecars written via .tmp + os.rename (atomic on crash)
  applied to step checkpoints, best.safetensors, and best.json

Documentation (15 fixes):
- CLAUDE.md: Qwen3 extraction layers 9,18,27 → 8,17,26 (C vs Python indexing);
  dispatch_queue.jsonl path fixed; flux_metal_sgemm → iris_metal_sgemm;
  dispatch queue visibility contradiction resolved
- DISPATCH.md: hardware 64 GB → 32 GB; dispatch_queue path fixed; Gap 2 marked
  resolved; validate_shards + training_warmup added to pipeline flow and sentinel
  list; heartbeat table completed (stager, mine, flywheel, ablation, trainer)
- README.md: step order corrected (build_shards before filter_shards); missing
  steps added
- BUGS.md: PIPE-1 shard block 250000 → 200000
- plans/: pipeline-v2 marked Implemented; lora-support marked Implemented;
  training-split-forward marked Superseded
- iris_shaders.metal: header filename flux_shaders → iris_shaders

+++ iris.c
@@ -1812,12 +1812,12 @@ iris_image *iris_img2img_debug_py(iris_ctx *ctx, const iris_params *params) {
         return NULL;
     }
     fseek(f_noise, 0, SEEK_END);
-    int noise_size = ftell(f_noise) / sizeof(float);
+    long noise_size = ftell(f_noise) / (long)sizeof(float);
     fseek(f_noise, 0, SEEK_SET);
     float *noise = (float *)malloc(noise_size * sizeof(float));
     fread(noise, sizeof(float), noise_size, f_noise);
     fclose(f_noise);
-    fprintf(stderr, "[DEBUG] Loaded noise: %d floats\n", noise_size);
+    fprintf(stderr, "[DEBUG] Loaded noise: %ld floats\n", noise_size);
 
     /* Load Python's ref_latent */
     FILE *f_ref = fopen("/tmp/py_ref_latent.bin", "rb");
